requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
altair=4.1.0=pypi_0
argon2-cffi=20.1.0=pypi_0
astor=0.8.1=pypi_0
async-generator=1.10=pypi_0
attrs=21.2.0=pypi_0
backcall=0.2.0=pypi_0
base58=2.1.0=pypi_0
bleach=3.3.0=pypi_0
blinker=1.4=pypi_0
ca-certificates=2021.5.25=h06a4308_1
cachetools=4.2.2=pypi_0
certifi=2021.5.30=py39h06a4308_0
cffi=1.14.5=pypi_0
chardet=4.0.0=pypi_0
click=7.1.2=pypi_0
decorator=5.0.9=pypi_0
defusedxml=0.7.1=pypi_0
entrypoints=0.3=pypi_0
filelock=3.0.12=pypi_0
gitdb=4.0.7=pypi_0
gitpython=3.1.18=pypi_0
htbuilder=0.5.0=pypi_0
huggingface-hub=0.0.12=pypi_0
idna=2.10=pypi_0
ipykernel=5.5.5=pypi_0
ipython=7.24.1=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.6.3=pypi_0
iteration-utilities=0.11.0=pypi_0
jedi=0.18.0=pypi_0
jinja2=3.0.1=pypi_0
joblib=1.0.1=pypi_0
jsonschema=3.2.0=pypi_0
jupyter-client=6.1.12=pypi_0
jupyter-core=4.7.1=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.0=pypi_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libstdcxx-ng=9.3.0=hd4cf53a_17
markupsafe=2.0.1=pypi_0
matplotlib-inline=0.1.2=pypi_0
mistune=0.8.4=pypi_0
nbclient=0.5.3=pypi_0
nbconvert=6.1.0=pypi_0
nbformat=5.1.3=pypi_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pypi_0
notebook=6.4.0=pypi_0
numpy=1.21.0=pypi_0
openssl=1.1.1k=h27cfd23_0
packaging=20.9=pypi_0
pandas=1.2.5=pypi_0
pandocfilters=1.4.3=pypi_0
parso=0.8.2=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.2.0=pypi_0
pip=21.1.2=py39h06a4308_0
prometheus-client=0.11.0=pypi_0
prompt-toolkit=3.0.19=pypi_0
protobuf=3.17.3=pypi_0
ptyprocess=0.7.0=pypi_0
pycparser=2.20=pypi_0
pydeck=0.6.2=pypi_0
pygments=2.9.0=pypi_0
pyparsing=2.4.7=pypi_0
pyrsistent=0.17.3=pypi_0
python=3.9.5=h12debd9_4
python-dateutil=2.8.1=pypi_0
pytz=2021.1=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=22.1.0=pypi_0
readline=8.1=h27cfd23_0
regex=2021.4.4=pypi_0
requests=2.25.1=pypi_0
sacremoses=0.0.45=pypi_0
send2trash=1.7.1=pypi_0
setuptools=52.0.0=py39h06a4308_0
six=1.16.0=pypi_0
smmap=4.0.0=pypi_0
sqlite=3.35.4=hdfb4753_0
st-annotated-text=1.1.0=pypi_0
streamlit=0.83.0=pypi_0
terminado=0.10.1=pypi_0
testpath=0.5.0=pypi_0
tk=8.6.10=hbc83047_0
tokenizers=0.10.3=pypi_0
toml=0.10.2=pypi_0
toolz=0.11.1=pypi_0
torch=1.9.0+cu111=pypi_0
torchaudio=0.9.0=pypi_0
torchvision=0.10.0+cu111=pypi_0
tornado=6.1=pypi_0
tqdm=4.61.1=pypi_0
traitlets=5.0.5=pypi_0
transformers=4.8.0=pypi_0
typing-extensions=3.10.0.0=pypi_0
tzdata=2020f=h52ac0ba_0
tzlocal=2.1=pypi_0
urllib3=1.26.5=pypi_0
validators=0.18.2=pypi_0
watchdog=2.1.2=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=pypi_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
